list files, including hidden ones, in a directory

list files, including hidden ones, in a directory

list files, including hidden ones, in a directory

list_files(path = ".", all.files = TRUE, recursive = TRUE, no.. = TRUE,
  ..., normalize = FALSE, must_work = TRUE)

list_files_and_dirs(...)

Arguments

path
a character vector of full path names; the default corresponds to the working directory, Default: '.'

all.files
a logical value. If FALSE, only the names of visible files are returned. If TRUE, all file names will be returned., Default: TRUE

recursive
logical. Should the listing recurse into directories?, Default: TRUE

no..
logical. Should both "." and ".." be excluded also from non-recursive listings?, Default: TRUE

...
args to pass to [dir()]

normalize
whether to convert file paths to canonical form for the platform, Default: FALSE

must_work
logical: if TRUE then an error is given if the result cannot be determined; if NA then a warning., Default: TRUE

...
parameters to pass to list_files

Value

vector of file paths

a list containing files and dirs elements